pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vector-explore"
version = "0.1.0"
description = "Local-first educational CLI for chunking, embeddings, vector search, and RAG."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Vector Explore" }]
dependencies = [
  "numpy>=1.26",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
st = [
  "sentence-transformers>=3.0.0",
]
lancedb = [
  "lancedb>=0.10.0",
]
qdrant = [
  "qdrant-client>=1.9.0",
]
pinecone = [
  "pinecone>=5.0.0",
]
cloud = [
  "httpx>=0.27.0",
]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
vector-explore = "vector_explore.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
